Make sure  of keyPair call returns the array of key pairs
This is in line with other calls like `giantSwarm.environments`.
It makes it so consumers of this client do not have to care what
the right key is to get to the data.

Currently the API returns the key pairs in a key called `KeyPairs`
The casing is not in line with the rest of the API. However
the js-client should shield implementors from this discrepancy.